relacionamento no entity

+++ src/main/java/br/com/fiap/webshift/model/CategoriaModel.java
@@ -19,7 +19,7 @@ public class CategoriaModel {
 	@Column(name = "ID_CATEGORIA")
 	@GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "CATEGORIA_SEQ")
 	@SequenceGenerator(name = "CATEGORIA_SEQ", initialValue = 1, allocationSize = 1)
-	private long idCategoria;
+	private int idCategoria;
 
 	@Column(name = "NOME_CATEGORIA", length = 50, nullable = false)
 	@NotNull(message = "Nome obrigatório")
@@ -30,16 +30,16 @@ public class CategoriaModel {
 		super();
 	}
 	
-	public CategoriaModel(long idCategoria, String nomeCategoria) {
+	public CategoriaModel(int idCategoria, String nomeCategoria) {
 		super();
 		this.idCategoria = idCategoria;
 		this.nomeCategoria = nomeCategoria;
 	}
 
-	public long getIdCategoria() {
+	public int getIdCategoria() {
 		return idCategoria;
 	}
-	public void setIdCategoria(long idCategoria) {
+	public void setIdCategoria(int idCategoria) {
 		this.idCategoria = idCategoria;
 	}
 	public String getNomeCategoria() {
